meme saving works fine

+++ docker/scripts/init.sql
@@ -1,6 +1,6 @@
 CREATE TABLE meme (
     meme_id serial PRIMARY KEY,
-    message_id bigint NOT NULL,
+    message_id integer NOT NULL,
     chat_id bigint NOT NULL,
     publish_date date NOT NULL,
     user_id bigint NOT NULL,
@@ -10,7 +10,7 @@ CREATE TABLE meme (
 CREATE TYPE score_types AS ENUM ('LIKE', 'DISLIKE', 'ACCORDION');
 
 CREATE TABLE meme_score (
-    meme_id bigint NOT NULL,
+    meme_id integer NOT NULL,
     user_id bigint NOT NULL,
     score score_types NOT NULL,
     PRIMARY KEY(meme_id, user_id),
